==> step2.4.doublet.scrublet.python.ipynb <==
%matplotlib inline
import scrublet as scr
import scipy.io
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import os


plt.rcParams['font.family'] = 'sans-serif'
plt.rcParams['font.sans-serif'] = 'Arial'
plt.rc('font', size=14)
plt.rcParams['pdf.fonttype'] = 42


input_dir = 'output2_step2.1.qc_scaledata/10xFormat'
counts_matrix = scipy.io.mmread(input_dir + '/matrix.mtx').T.tocsc()
genes = np.array(scr.load_genes(input_dir + '/genes.tsv', delimiter='\t', column=1))

print('Counts matrix shape: {} rows, {} columns'.format(counts_matrix.shape[0], counts_matrix.shape[1]))
print('Number of genes in gene list: {}'.format(len(genes)))

import scanpy as sc

adata=sc.read('0_convertSeuratTOh5ad/TotalTissue.combined.qc.filtered.h5ad')
adata

sampleid = adata.obs.Sample.unique().tolist()
adata_list= []
for id in sampleid:
    adata_subset = adata[adata.obs['Sample'].isin([id])]
    scrub = scr.Scrublet(adata_subset.X, expected_doublet_rate=0.06,sim_doublet_ratio = 2)
    doublet_scores, predicted_doublets = scrub.scrub_doublets(min_counts=2, 
                                                          min_cells=3, 
                                                          min_gene_variability_pctl=85, 
                                                          n_prin_comps=30)
    adata_subset.obs["scrubletScore"] = scrub.doublet_scores_obs_
    adata_subset.obs["scrublet_predict"] = scrub.predicted_doublets_
    adata_list.append(adata_subset.obs)
    

adata_list_merged = pd.concat(adata_list)

adata_list_merged.to_csv('output2_step2.4.doublet.scrublet.python/mergedadatalistScrubletOutput_default.csv')

sampleid = adata.obs.Sample.unique().tolist()
adata_list= []
for id in sampleid:
    adata_subset = adata[adata.obs['Sample'].isin([id])]
    scrub = scr.Scrublet(adata_subset.X, expected_doublet_rate=0.06,sim_doublet_ratio = 2,n_neighbors=30)
    doublet_scores, predicted_doublets = scrub.scrub_doublets(min_counts=2, 
                                                          min_cells=3, 
                                                          min_gene_variability_pctl=85, 
                                                          n_prin_comps=30)
    adata_subset.obs["scrubletScore"] = scrub.doublet_scores_obs_
    adata_subset.obs["scrublet_predict"] = scrub.predicted_doublets_
    adata_list.append(adata_subset.obs)
adata_list_merged = pd.concat(adata_list)
adata_list_merged.to_csv('output2_step2.4.doublet.scrublet.python/mergedadatalistScrubletOutput_neighbors30.csv')
